# Improve report performance with high-cardinality import joins