--- a/lib/plausible/stats/sql/special_metrics.ex
+++ b/lib/plausible/stats/sql/special_metrics.ex
@@ -26,7 +26,8 @@ defmodule Plausible.Stats.SQL.SpecialMetrics do
         |> remove_filters_ignored_in_totals_query()
         |> Query.set(
           dimensions: [],
-          include_imported: query.include_imported
+          include_imported: query.include_imported,
+          pagination: nil
         )
 
       q
@@ -57,7 +58,11 @@ defmodule Plausible.Stats.SQL.SpecialMetrics do
         |> Query.set(
           dimensions: [],
           include_imported: query.include_imported,
+<<<<<<< HEAD
           preloaded_goals: []
+=======
+          pagination: nil
+>>>>>>> 71ef81755 (Splitting should no longer remove pagination. Handle special cases in special_metrics.ex)
         )
 
       q
@@ -100,7 +105,11 @@ defmodule Plausible.Stats.SQL.SpecialMetrics do
           metrics: [:visitors],
           order_by: [],
           include_imported: query.include_imported,
+<<<<<<< HEAD
           preloaded_goals: []
+=======
+          pagination: nil
+>>>>>>> 71ef81755 (Splitting should no longer remove pagination. Handle special cases in special_metrics.ex)
         )
 
       from(e in subquery(q),