@macobo 2024-11-25
Marking as a draft for a bit. Need to figure out if there's cases where the removed splitting-removes-pagination logic would break things that tests aren't covering.